requirements.txt
beautifulsoup4==4.11.2
fastapi==0.91.0
jaro_winkler==2.0.3
jose==1.0.0
openpyxl==3.0.10
pandas==1.5.2
passlib==1.7.4
pydantic==1.10.4
python_jose==3.3.0
requests==2.28.1
SQLAlchemy==1.4.46
